Moved message group b y to function
