+++ packages/application-audit-archiver/src/handler/messagesHandler.ts
@@ -1,20 +1,14 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
-/* eslint-disable functional/immutable-data */
 import { KafkaMessage } from "kafkajs";
 import {
   FileManager,
   Logger,
-  decodeKafkaMessage,
   formatTimehhmmss,
 } from "pagopa-interop-kpi-commons";
-import {
-  ApplicationAuditEvent,
-  generateId,
-  genericInternalError,
-} from "pagopa-interop-kpi-models";
+import { generateId, genericInternalError } from "pagopa-interop-kpi-models";
 import { config } from "../config/config.js";
 import { compressJson } from "../utilities/compression.js";
+import { groupMessagesByDate } from "../utilities/groupMessagesByDate.js";
 
 export async function handleMessages(
   messages: KafkaMessage[],
@@ -22,23 +16,9 @@ export async function handleMessages(
   logger: Logger
 ) {
   try {
-    const grouped: Record<string, ApplicationAuditEvent[]> = {};
+    const groupedMessages = groupMessagesByDate(messages);
 
-    for (const message of messages) {
-      const item = decodeKafkaMessage(message, ApplicationAuditEvent);
-      const date = new Date(item.timestamp);
-      const year = date.getUTCFullYear();
-      const month = String(date.getUTCMonth() + 1).padStart(2, "0");
-      const day = String(date.getUTCDate()).padStart(2, "0");
-
-      const key = `${year}-${month}-${day}`;
-      if (!grouped[key]) {
-        grouped[key] = [];
-      }
-      grouped[key].push(item);
-    }
-
-    for (const [dateKey, groupMessages] of Object.entries(grouped)) {
+    for (const [dateKey, groupMessages] of Object.entries(groupedMessages)) {
       const [year, month, day] = dateKey.split("-");
       const jsonString = JSON.stringify(groupMessages);
       const compressedBuffer = await compressJson(jsonString);
